+++ SwiftySCAD/SwiftySCADFramework/Constants.swift
@@ -4,6 +4,6 @@
 
 import Foundation
 
-let epsilon: Float = 0.001
-let epsilon_2: Float = epsilon/2.0
-let big_epsilon: Float = 1
+let epsilon: Double = 0.001
+let epsilon_2: Double = epsilon/2.0
+let big_epsilon: Double = 1
