+++ SwiftySCAD/SwiftySCADFramework/CustomTypes.swift
@@ -7,6 +7,6 @@ import Foundation
 
 
 struct Size {
-    var width: Float
-    var height: Float
+    var width: Double
+    var height: Double
 }
